FileSorting - commit 17

+++ FileSorting.Generator/GeneratorManager.cs
@@ -10,18 +10,22 @@ namespace FileSorting.Generator
 {
     public class GeneratorManager
     {
+        #region Members
         private readonly Random random;
         private readonly char[] letters;
-        private readonly char[] digits;      
+        private readonly char[] digits;
+        #endregion
 
+        #region Constructor
         public GeneratorManager()
         {
             random = new Random();
-            //letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ ".ToCharArray();
-            letters = "abcdefghijklmnopqrstuvwxyz".ToCharArray();
+            letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ ".ToCharArray();
             digits = "0123456789".ToCharArray();
         }
+        #endregion
 
+        #region Generators
         public async Task<(bool, string)> GenerateFileAsync(string sourceFilePath, double fileSizeMB)
         {
             var result = await GenerateFileInternalAsync(sourceFilePath, fileSizeMB);
@@ -36,27 +40,37 @@ namespace FileSorting.Generator
 
         private async Task<(bool, string)> GenerateFileInternalAsync(string sourceFilePath, double fileSizeMB)
         {
-            var folderPath = Path.GetDirectoryName(sourceFilePath);
-            var dividedFilesDirectory = Path.Combine(folderPath, Consts.FILE_SORTING_TEMP_FOLDER);
-            var tasks = new List<Task>();
-
-            if (Directory.Exists(dividedFilesDirectory))
-                Directory.Delete(dividedFilesDirectory, true);
-
-            Directory.CreateDirectory(dividedFilesDirectory);
-
-            for (int i = 1; i <= Consts.NUMBER_OF_PARALLEL_WRITERS; i++)
+            try
             {
-                var task = Task.Run(async () =>
-                {
-                    var outputFilePath = Path.Combine(dividedFilesDirectory, $"output_{Guid.NewGuid()}.txt");
-                    await GenerateFileSingleAsync(outputFilePath, fileSizeMB / Consts.NUMBER_OF_PARALLEL_WRITERS);
-                });
-                tasks.Add(task);
-            }
-            await Task.WhenAll(tasks.ToArray());
+                var folderPath = Path.GetDirectoryName(sourceFilePath);
+                if (string.IsNullOrEmpty(folderPath))
+                    return (false, "folderPath is null or empty");
 
-            return (true, string.Empty);
+                var dividedFilesDirectory = Path.Combine(folderPath, Consts.FILE_SORTING_TEMP_FOLDER);
+                var tasks = new List<Task>();
+
+                if (Directory.Exists(dividedFilesDirectory))
+                    Directory.Delete(dividedFilesDirectory, true);
+
+                Directory.CreateDirectory(dividedFilesDirectory);
+
+                for (int writerIndex = 1; writerIndex <= Consts.NUMBER_OF_PARALLEL_WRITERS; writerIndex++)
+                {
+                    var task = Task.Run(async () =>
+                    {
+                        var outputFilePath = Path.Combine(dividedFilesDirectory, $"output_{Guid.NewGuid()}.txt");
+                        await GenerateFileSingleAsync(outputFilePath, fileSizeMB / Consts.NUMBER_OF_PARALLEL_WRITERS);
+                    });
+                    tasks.Add(task);
+                }
+                await Task.WhenAll(tasks.ToArray());
+
+                return (true, string.Empty);
+            }
+            catch(Exception ex)
+            {
+                return (false, ex.Message);
+            }
         }
 
         private async Task<(bool,string)> GenerateFileSingleAsync(string filePath, double fileSizeMB)
@@ -80,10 +94,10 @@ namespace FileSorting.Generator
                             if (blockSizeMB + resultGetFileSize.Item2 > fileSizeMB)
                             {
                                 var subBlock = new List<string>();
-                                for (int i = 0; i < blocks.Count; i += Consts.SUB_BLOCK_STEP)
+                                for (int blockStep1 = 0; blockStep1 < blocks.Count; blockStep1 += Consts.SUB_BLOCK_STEP)
                                 {
-                                    for (int j = i; j < i + Consts.SUB_BLOCK_STEP && j < blocks.Count; j++)
-                                        subBlock.Add(blocks[j]);
+                                    for (int blockStep2 = blockStep1; blockStep2 < blockStep1 + Consts.SUB_BLOCK_STEP && blockStep2 < blocks.Count; blockStep2++)
+                                        subBlock.Add(blocks[blockStep2]);
 
                                     blockSizeMB = GetBlockSizeMB(subBlock);
                                     if (blockSizeMB + resultGetFileSize.Item2 > fileSizeMB)                                    
@@ -110,29 +124,6 @@ namespace FileSorting.Generator
             }           
         }
 
-        private (bool,double,string) GetFileSize(string filePath)
-        {
-            try
-            {
-                var fileInfo = new FileInfo(filePath);
-                if (fileInfo.Exists)
-                {
-                    long fileSizeInBytes = fileInfo.Length;
-                    double fileSizeInKilobytes = fileSizeInBytes / Consts.KILOBYTE_SIZE;
-                    double fileSizeInMegabytes = fileSizeInKilobytes / Consts.KILOBYTE_SIZE;
-                    return (true, fileSizeInMegabytes, string.Empty);
-                }
-                else
-                {
-                    return (true, 0, string.Empty);
-                }
-            }
-            catch (Exception ex)
-            {
-                return (false, 0, ex.ToString());
-            }
-        }
-
         private List<string> GenerateFileBlockSingle()
         {
             var stringList = new List<string>();
@@ -144,7 +135,7 @@ namespace FileSorting.Generator
                 var numberOfRepetions = random.Next(1, Consts.REPETIONS_MAX);
                 var randomString = GenerateRandomString(lengthLetters);
 
-                for (int i = 0; i < numberOfRepetions && linesCounter < Consts.WRITE_BLOCK_SIZE; i++)
+                for (int repetition = 0; repetition < numberOfRepetions && linesCounter < Consts.WRITE_BLOCK_SIZE; repetition++)
                 {
                     var randomNumber = GenerateRandomNumber(lengthDigits);
                     var randomCombinedString = $"{randomNumber}. {randomString}";
@@ -155,61 +146,34 @@ namespace FileSorting.Generator
 
             return stringList
             .OrderBy(line => Guid.NewGuid())
-            .ToList();                      
+            .ToList();
         }
 
         private async Task<ConcurrentBag<string>> GenerateFileBlockMultipleAsync(int numberOfBlocks)
         {
             var tasks = new List<Task>();
             var allBlocks = new ConcurrentBag<string>();
-            
-            for (int i = 0; i < numberOfBlocks; i++)
+
+            for (int block = 0; block < numberOfBlocks; block++)
             {
                 var task = Task.Run(() =>
                 {
-                    foreach(var line in GenerateFileBlockSingle())
+                    foreach (var line in GenerateFileBlockSingle())
                         allBlocks.Add(line);
                 });
-                tasks.Add(task);          
+                tasks.Add(task);
             }
 
             await Task.WhenAll(tasks.ToArray());
             return allBlocks;
         }
 
-        private double GetBlockSizeMB(IEnumerable<string> block)
-        {
-            var bytes = new List<byte>();
-            foreach (var line in block)
-            {
-                bytes.AddRange(Encoding.UTF8.GetBytes(line));
-            }
-
-            var sizeInMB = (double)bytes.ToArray().Length / (Consts.KILOBYTE_SIZE * Consts.KILOBYTE_SIZE);
-            return sizeInMB;         
-        }
-
-        private async Task<(bool,string)> WriteBlockToFileAsync(string filePath, IEnumerable<string> block)
-        {
-            try
-            {
-                await File.AppendAllLinesAsync(filePath, block);
-                return (true, string.Empty);
-            }
-            catch (Exception ex)
-            {
-                return (false, ex.ToString());
-            }
-        }
-     
         private string GenerateRandomString(int length)
         {
             var builder = new StringBuilder();
 
-            for (int i = 0; i < length; i++)
-            {
+            for (int charIndex = 0; charIndex < length; charIndex++)
                 builder.Append(letters[random.Next(letters.Length)]);
-            }
 
             return builder.ToString();
         }
@@ -218,21 +182,25 @@ namespace FileSorting.Generator
         {
             var builder = new StringBuilder();
 
-            for (int i = 0; i < length; i++)
-            {
+            for (int charIndex = 0; charIndex < length; charIndex++)
                 builder.Append(digits[random.Next(digits.Length)]);
-            }
 
             return builder.ToString();
         }
+        #endregion
 
+        #region Mergers
         private Task<(bool, string)> MergeDividedFiles(string sourceFilePath)
         {
             return Task.Run(async () =>
             {
                 try
                 {
-                    var dividedFilesDirectory = Path.Combine(Path.GetDirectoryName(sourceFilePath), Consts.FILE_SORTING_TEMP_FOLDER);
+                    var directoryName = Path.GetDirectoryName(sourceFilePath);
+                    if (string.IsNullOrEmpty(directoryName))
+                        return (false, "directoryName is null or empty");
+
+                    var dividedFilesDirectory = Path.Combine(directoryName, Consts.FILE_SORTING_TEMP_FOLDER);
                     var tasks = new List<Task>();
                     var bufferCounter = 0;
 
@@ -242,24 +210,24 @@ namespace FileSorting.Generator
                         if (files.Length == 1)
                             break;
 
-                        for (int i = 0; i < files.Length; i += 2)
+                        for (int fileIndex = 0; fileIndex < files.Length; fileIndex += 2)
                         {
                             if (bufferCounter < Consts.NUMBER_OF_PARALLEL_READERS &&
-                                i < files.Length - 1)
+                                fileIndex < files.Length - 1)
                             {
-                                int capturedI = i;
-                                int capturedIplus1 = i + 1;
+                                int capturedFileIndexI = fileIndex;
+                                int capturedFileIndexIplus1 = fileIndex + 1;
                                 tasks.Add(Task.Run(() =>
                                 {
-                                    MergeDividedFiles(files[capturedI], files[capturedIplus1], Path.Combine(dividedFilesDirectory, $"merged_{Guid.NewGuid()}.txt"));
-                                    File.Delete(files[capturedI]);
-                                    File.Delete(files[capturedIplus1]);
+                                    MergeDividedFiles(files[capturedFileIndexI], files[capturedFileIndexIplus1], Path.Combine(dividedFilesDirectory, $"merged_{Guid.NewGuid()}.txt"));
+                                    File.Delete(files[capturedFileIndexI]);
+                                    File.Delete(files[capturedFileIndexIplus1]);
                                 }));
                                 bufferCounter++;
                             }
 
                             if (bufferCounter == Consts.NUMBER_OF_PARALLEL_READERS ||
-                                i + 2 >= files.Length - 1)
+                                fileIndex + 2 >= files.Length - 1)
                             {
                                 await Task.WhenAll(tasks.ToArray());
                                 tasks.Clear();
@@ -283,8 +251,8 @@ namespace FileSorting.Generator
             using (var reader2 = new StreamReader(file2Path))
             using (var writer = new StreamWriter(outputFilePath))
             {
-                string line1 = reader1.ReadLine();
-                string line2 = reader2.ReadLine();
+                var line1 = reader1.ReadLine();
+                var line2 = reader2.ReadLine();
 
                 while (line1 != null || line2 != null)
                 {
@@ -299,20 +267,29 @@ namespace FileSorting.Generator
                 }
             }
         }
+        #endregion
 
+        #region Helpers
         private Task<(bool, string)> MoveGeneratedFileToSourceLocation(string sourceFilePath)
         {
             return Task.Run(() =>
             {
                 try
                 {
-                    var sourceDirectory = Path.Combine(Path.GetDirectoryName(sourceFilePath), Consts.FILE_SORTING_TEMP_FOLDER);
                     var targetDirectory = Path.GetDirectoryName(sourceFilePath);
+                    if (string.IsNullOrEmpty(targetDirectory))
+                        return (false, "directoryName is null or empty");
+
+                    var sourceDirectory = Path.Combine(targetDirectory, Consts.FILE_SORTING_TEMP_FOLDER);
+
                     var files = Directory.GetFiles(sourceDirectory);
+                    if (files == null || !files.Any())
+                        return (false, "There is no files in sourceDirectory");
 
                     var newFileName = $"{Path.GetFileNameWithoutExtension(sourceFilePath)}.txt";
                     var targetFilePath = Path.Combine(targetDirectory, newFileName);
-                    File.Move(files.FirstOrDefault(), targetFilePath);
+
+                    File.Move(files.First(), targetFilePath);
                     Directory.Delete(sourceDirectory);
 
                     return (true, string.Empty);
@@ -323,5 +300,52 @@ namespace FileSorting.Generator
                 }
             });
         }
+
+        private async Task<(bool, string)> WriteBlockToFileAsync(string filePath, IEnumerable<string> block)
+        {
+            try
+            {
+                await File.AppendAllLinesAsync(filePath, block);
+                return (true, string.Empty);
+            }
+            catch (Exception ex)
+            {
+                return (false, ex.ToString());
+            }
+        }
+
+        private (bool, double, string) GetFileSize(string filePath)
+        {
+            try
+            {
+                var fileInfo = new FileInfo(filePath);
+                if (fileInfo.Exists)
+                {
+                    long fileSizeInBytes = fileInfo.Length;
+                    double fileSizeInKilobytes = fileSizeInBytes / Consts.KILOBYTE_SIZE;
+                    double fileSizeInMegabytes = fileSizeInKilobytes / Consts.KILOBYTE_SIZE;
+                    return (true, fileSizeInMegabytes, string.Empty);
+                }
+                else
+                {
+                    return (true, 0, string.Empty);
+                }
+            }
+            catch (Exception ex)
+            {
+                return (false, 0, ex.ToString());
+            }
+        }
+
+        private double GetBlockSizeMB(IEnumerable<string> block)
+        {
+            var bytes = new List<byte>();
+            foreach (var line in block)          
+                bytes.AddRange(Encoding.UTF8.GetBytes(line));
+            
+            var sizeInMB = (double)bytes.ToArray().Length / (Consts.KILOBYTE_SIZE * Consts.KILOBYTE_SIZE);
+            return sizeInMB;
+        }
+        #endregion
     }
 }
